+++ src/Regex/Validation.java
@@ -1,3 +1,5 @@
+package Regex;
+
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -22,7 +24,8 @@ public class Validation {
     }
     public void emailValidation(String email)
     {
-        String regex = "^[abc]{1}(.*[A-Za-z0-9-_+.])?@{1}[A-Za-z0-9]{1,}.com{1}(.[a-z]*)?$";
+//        String regex = "^[abc]{1}(.*[A-Za-z0-9-_+.])?@{1}[A-Za-z0-9]{1,}.com{1}(.[a-z]*)?$";
+        String regex = "^[abc]{1}(.*[A-Za-z0-9-_+])(?(?=.*.)+[A-Za-z0-9-_+]{1,})?@{1}[A-Za-z0-9]{1,}.com{1}(.[a-z]*)?$";
         Pattern p = Pattern.compile(regex);
         Matcher m = p.matcher(email);
         check(m.find());
@@ -37,7 +40,7 @@ public class Validation {
     public void passwordValidation(String password)
     {
         //This is for rule 4
-        String regex = "^(?=.*[A-Z])(?=.*[0-9])(?=.{1}?[@#$%^&-+=()])([a-zA-Z0-9]*).{8,}$";
+        String regex = "^(?=.*[A-Z])(?=.*[0-9])(?=.{1}?[@#$%^&-+=()])(?=.*[a-zA-Z0-9]*).{8,}$";
         Pattern p = Pattern.compile(regex);
         Matcher m = p.matcher(password);
         check(m.find());
